% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df_to_files.R
\name{split_df_to_files}
\alias{split_df_to_files}
\title{Dataframe is splitted by one or more variable and depending of selected mode exported to xlsx file with group named tabs or to multiple csv files.}
\usage{
split_df_to_files(df, file_type, file_name, ...)
}
\arguments{
\item{df}{Data frame}

\item{file_type}{export mode (csv or xlsx)}

\item{file_name}{file name}

\item{...}{grouping variable}
}
\value{
file or files
}
\description{
Dataframe is splitted by one or more variable and depending of selected mode exported to xlsx file with group named tabs or to multiple csv files.
}
